refactor: adaptation after cherrypick
