+++ src/app/signup/page.tsx
@@ -2,6 +2,7 @@ import Link from "next/link";
 import Image from "next/image";
 
 import {SignupForm} from "@/components";
+import {aladin} from "@/lib/utils";
 
 import styles from "./signup.module.css";
 
@@ -22,7 +23,7 @@ const SignupPage = () => {
             <div className={styles.rightColumn}>
                 <div className={styles.logoContainer}>
                     <Link href="/" className={styles.logo}>
-                        <h1 className={styles.logoText}>Project RPG</h1>
+                        <h1 className={`${styles.logoText} ${aladin.className}`}>Project RPG</h1>
                     </Link>
                 </div>
