+++ src/block/index.ts
@@ -1,5 +1,7 @@
 import {Header, Footer, Hero, Presentation, CampaignDemo, UsecaseDemo, WorldDemo} from "@/block/Home";
+import Campaigns from "@/block/Campaigns/Campaigns";
 
 export {
-    Header, Footer, Hero, Presentation, CampaignDemo, UsecaseDemo, WorldDemo
+    Header, Footer, Hero, Presentation, CampaignDemo, UsecaseDemo, WorldDemo,
+    Campaigns
 };
